Migrating to go-fpdf/fpdf@v0.4.0 improves somewhat plotting performances:

```
  name           old time/op    new time/op    delta
  Canvas-8          221µs ± 2%     181µs ± 4%  -18.25%  (p=0.000 n=30+29)
  CanvasImage-8    72.0ms ± 4%    72.3ms ± 4%     ~     (p=0.406 n=27+29)

  name           old alloc/op   new alloc/op   delta
  Canvas-8          111kB ± 1%     109kB ± 1%   -1.79%  (p=0.000 n=27+23)
  CanvasImage-8    2.94MB ± 0%    2.69MB ± 0%   -8.64%  (p=0.000 n=29+29)

  name           old allocs/op  new allocs/op  delta
  Canvas-8          1.48k ± 0%     0.87k ± 0%  -41.09%  (p=0.000 n=30+30)
  CanvasImage-8       121 ± 0%        84 ± 0%  -30.58%  (p=0.000 n=26+27)
```

Fixes #697.

Please take a look.

<!--
Checklist:

- API changes have been discussed
- code is goformated correctly (goimports)
- packages with generated code have had code generation run
- tests pass locally
- linked to relevant issues

Please make sure your commit message summary line and pull request
title match the Go convention; a one-line summary of the change,
prefixed by the primary affected package that should complete the
sentence, "This change modifies Gonum to _____."
-->